SITCA File Transfer : Perf fix with Random Access thread safe read + Fix for Rounding of int sized long values resulting in negative offset

+++ FilePartsData.cs
@@ -29,9 +29,9 @@ namespace SITCAFileTransferService
 
         public string fileName;
 
-        public int currentOffset;
+        public long currentOffset;
 
-        public int currentIterationCount;
+        public long currentIterationCount;
 
         public IMongoCollection<FilePartsData> currentCollection;
